+++ cmd/evolution-go/main.go
@@ -90,15 +90,15 @@ type chatwootMessageSenderAdapter struct {
 
 // Devolvem o ID da mensagem no WhatsApp pro chatwoot_service marcar o envio do
 // agente e ignorar o eco fromMe correspondente (anti-loop).
-func (a chatwootMessageSenderAdapter) SendText(number, text string, instance *instance_model.Instance) (string, error) {
+func (a chatwootMessageSenderAdapter) SendText(number, text string, instance *instance_model.Instance) (string, string, error) {
 	resp, err := a.sendService.SendText(&send_service.TextStruct{Number: number, Text: text}, instance)
 	if err != nil || resp == nil {
-		return "", err
+		return "", "", err
 	}
-	return resp.Info.ID, nil
+	return resp.Info.ID, resp.Info.Chat.String(), nil
 }
 
-func (a chatwootMessageSenderAdapter) SendMedia(number string, data []byte, mediaType, filename, caption string, instance *instance_model.Instance) (string, error) {
+func (a chatwootMessageSenderAdapter) SendMedia(number string, data []byte, mediaType, filename, caption string, instance *instance_model.Instance) (string, string, error) {
 	resp, err := a.sendService.SendMediaFile(&send_service.MediaStruct{
 		Number:   number,
 		Type:     mediaType,
@@ -106,9 +106,9 @@ func (a chatwootMessageSenderAdapter) SendMedia(number string, data []byte, medi
 		Filename: filename,
 	}, data, instance)
 	if err != nil || resp == nil {
-		return "", err
+		return "", "", err
 	}
-	return resp.Info.ID, nil
+	return resp.Info.ID, resp.Info.Chat.String(), nil
 }
 
 func init() {